move summing vars method to helpers file
